feat: better helpers
