+++ src/action.ts
@@ -1,8 +1,7 @@
 import * as core from '@actions/core'
 import * as github from '@actions/github'
 
-import {GitHubFileReader} from './utils/fileReader'
-import {getMonoRepo, validateSanityTemplate} from './utils/validator'
+import {validateRemoteTemplate} from './remote'
 
 async function run(): Promise<void> {
   try {
@@ -10,10 +9,7 @@ async function run(): Promise<void> {
     const branch = github.context.ref.replace('refs/heads/', '')
     const directory = core.getInput('directory', {required: false}) || ''
     const baseUrl = `https://raw.githubusercontent.com/${owner}/${repo}/${branch}/${directory}`
-
-    const fileReader = new GitHubFileReader(baseUrl)
-    const packages = (await getMonoRepo(fileReader)) || ['']
-    const result = await validateSanityTemplate(fileReader, packages)
+    const result = await validateRemoteTemplate(baseUrl)
 
     if (!result.isValid) {
       core.setFailed(result.errors.join('\n'))
