+++ src/index.ts
@@ -1,6 +1,6 @@
-export {validateLocal} from './local'
+export {validateLocalTemplate} from './local'
+export {validateRemoteTemplate} from './remote'
 export {ENV_FILE, ENV_TEMPLATE_FILES, REQUIRED_ENV_VAR} from './utils/constants'
-export type {FileReader} from './utils/fileReader'
-export {GitHubFileReader, LocalFileReader} from './utils/fileReader'
+export {type FileReader, GitHubFileReader, LocalFileReader} from './utils/fileReader'
 export type {ValidationResult} from './utils/types'
-export {getMonoRepo, validateSanityTemplate} from './utils/validator'
+export {getMonoRepo, validateTemplate} from './utils/validator'
